Implement endpoint for searching destinations by name.
The endpoint can be, for example:

```
/destinos?nome=Paris
/destinos?name=paris
```

Note that the search field can be 'name' or 'nome' and will look for
exact, case-insensitive matches.

If no match is found, the HTTP 404 status is returned with the response

```json
{
    "message": "Nenhum destino encontrado."
}
```

The filter implementation requires the 'django-filter' package.
The virtual environment requirements (`requirements.txt`) have been
updated accordingly.