Refactor Destination model and its CRUD.
The `photo` field is now a proper static image, which will be stored in
`/media_root` and accessible bia de url `/media/`.

The `price` field is now 'MoneyField' implemented with the
'django-moneu' package.
The default currency is set to Brazilian Real.

The virtual environment requirements are also updated accordingly, as
the package 'django-money' is required now.